requirements.txt
coverage==4.5.2
Django==2.1.3
djangorestframework==3.9.0
Pillow==5.3.0
pytz==2018.7
